services/nff-data-ingestion/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# HTTP clients
httpx==0.26.0
requests==2.31.0

# Database
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis
redis==5.0.1

# Data processing
pandas==2.1.4
numpy==1.26.2
openpyxl==3.1.2

# Configuration
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Logging and monitoring
sentry-sdk==1.39.2

# Scheduler for cron jobs
APScheduler==3.10.4

# Date/time
python-dateutil==2.8.2

# Excel reading
xlrd==2.0.1

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# AI/ML
openai==1.12.0

# Linting (optional)
black==23.12.1
flake8==7.0.0
mypy==1.8.0